List of figures, tables and bibliography title fix
ixed optional cutom titles for the LoF, LoT and for the bibliography.  The previous commands used for this only affected the table of contents, but they happened to have the same values as default, thus hiding this error.  I have added some commented lines in both languages that show how to use custom titles, if needed.